Go back to old test partitioning
